Bug 377552 -Accessing session from shared plugin not possible
https://bugs.eclipse.org/bugs/show_bug.cgi?id=377552

The behaviour of the methods in ClientSessionRegistryService was confusing: They had the same name (newClientSesison) but one started the session in a new thread, the other one not. 
Fixed by harmonizing the behaviour: The session is now always started in the caller thread, no new thread is generated. 

Following threads are used to startup the session:
- Swing: Application (Main) thread
- Swt: Custom Thread (Job)
- Rwt: ServletJob

Important: It's never the UI-Job itself.