Bug 378389 -Rap: IForm.isCacheBounds not supported so far
https://bugs.eclipse.org/bugs/show_bug.cgi?id=378389

FormBoundsProvider moved from RwtScoutDesktopForm to RwtScoutViewStack to make sure the effective width is stored which is the one of the viewStack and not the one of the form. Also added a resize listener to make sure the bounds are always stored even when closing the browser (which does not lead to a proper logout).