https://bugs.eclipse.org/bugs/show_bug.cgi?id=379427: "SWT: Every second startup the client closes all views when running with Eclipse 4.2"
SWT: when the workbench is started, it restores the views that have been open when closing it the last time. As the application should always start with the defined views, these have been closed again when the SwtEnvironment is initialized. This code was implemented two times: once when initializing the environment and once when the gui is attached. this leads in some cases to the behavior that too many views are closed because the hide is called twice. There this change removes one of them.
Change discussed, reviewed and tested with Claudio Guglielmo.