[Bug 378151] Rap: Support for mobile devices
https://bugs.eclipse.org/bugs/show_bug.cgi?id=378151

Forms with display hint dialog are now properly displayed as dialogs and not as view anymore. Big advantage is that an overlay in the back is possible to make clear the user has to edit the form. 
This also improves performance on mobile version because the form in the back does not need to be created again after closing the top form.