[Bug 378151] Rap: Support for mobile devices
https://bugs.eclipse.org/bugs/show_bug.cgi?id=378151

Device navigation optimized: Back button now always goes one "drill down" back. Until now the regular navigation history has been used with the problem that bookmarks were also created after a search, which made the back button confusing. To solve that INavigationPoint has been introduced. This enables building a proper history which is also necessary for a possible bread crumbs support.
Also removed FormStack which was only necessary to always have one active form at the same time. This can be done far more easy in the gui with the benefit, that the real form list on the desktop does not need to be modified.