[Bug 378151] Rap: Support for mobile devices
https://bugs.eclipse.org/bugs/show_bug.cgi?id=378151

Removed form bounds caching form mobile desktop forms. It's mainly done because the right view stack shows forms with different width which should automatically be set if the form gets displayed. If the formbounds were stored which is done on resizing the width of the last shown form would be used and not always the one of the currently shown one.