[Bug 383774] Rap: Shortening behavior of CLabelEx too static, ellipsis does not dissapear if the label gets more space
https://bugs.eclipse.org/bugs/show_bug.cgi?id=383774

At the very first attempt to short the text (after server start), TextSizeUtil seems to return a wrong length. It is too big. So the first label to be shortened will get too small. For example TextSizeUtil returns 19 instead of 9 for the ellipsis. Prior to the addition of the resize listener setBounds has been used to update the text, which is called several times. Somehow the last time it is called the TextSizeUtil returns the correct value. That's why I reverted the deletion of this method.