Bug 364449 -Configuring Fields invisible on certain Devices
https://bugs.eclipse.org/bugs/show_bug.cgi?id=364449

Problem: ClientJob.getCurrentSession() and ClientSessionRegistryService#newClientSession are not sufficient. Outside a client job in a plugin with no dependency to the ui plugin it is not possible to get the client session, if it has not explicitly been  saved.
Fixed by filling the cache in the ClientSessionRegistryService and added a new method getClientSessionFor(). The existing method getClientSession() has not been used for this purpose to keep the deprecation warning.