[Bug 378151] Rap: Support for mobile devices
https://bugs.eclipse.org/bugs/show_bug.cgi?id=378151

Refactored drill down capability: It's no longer necessary to click on the drill down button. Drill down can now be executed by clicking on the row itself. This made it necessary to move the context menu buttons to the detail form. If there is no detail form it automatically will create one out of the table row.