[Bug 378151] Rap: Support for mobile devices
https://bugs.eclipse.org/bugs/show_bug.cgi?id=378151

Fixed issue, that editing or removing and element in the page form table did not refresh the current view. 
The fix in PageForm#formAddedNotify is essential: While the table data is getting replaced the child pages are still dirty. During this time the data must not be loaded again which would happen by selecting the page. The fix also improves performance.