[Bug 378151] Rap: Support for mobile devices
https://bugs.eclipse.org/bugs/show_bug.cgi?id=378151

Smartfield enhancement. 
Use case: The smartfield has a "new" menu which creates new smartfield objects. After creating such an object the app typically sets the value into the smartfield so that the user does not need to search for the object again. To make this work on the mobile client the value is automatically inserted in the filter field if it gets changed by such an external action.