[Bug 378151] Rap: Support for mobile devices
https://bugs.eclipse.org/bugs/show_bug.cgi?id=378151

Fixed bookmark activation: If a page got activated on the tablet client which is supposed to be on the right side, it was shown on the left side instead. Fixed by adding a better computation of the page slot. This also required an automatic opening of the left side page when only the right one gets selected.