[Bug 396025] Rap: Reload of the page disposes display but doesn't clean up the environment
https://bugs.eclipse.org/bugs/show_bug.cgi?id=396025

RwtScoutNavigationSupport does still sometimes not remove the listener because the job is not executed. Also can't solve the issue with join(5000), because rap locks the device on shutdown which prevents the JobChangeListeners from being executed -> deadlock.
With runNow it's guaranteed that the code will be executed, but actually by the wrong thread. But this shouldn't be any problem here.