Bug 387360 - RAP: When leaving the Rap application and going back a new session is started
https://bugs.eclipse.org/bugs/show_bug.cgi?id=387360

Bound the clientSession to the http session -> clientSession can be reused on page reload. Main difference in behaviour: Rap session along with the Display and RwtEnvironment get destroyed, but the scout model not. Therefore it's important that the gui properly detaches from the model, so that the memory can be cleaned up.
ClientSession gets destroyed on http session invalidation, which happens on session timeout or logout.