Bug 397991 - Rap/SWT: Issues with custom text smartfield
https://bugs.eclipse.org/bugs/show_bug.cgi?id=397991

Model was not updated because input verifier was not called -> TraverseEvent is only fired on TAB, therefore a FocusOutListener is required. SwtScoutComposite already takes care of this case: It calls handleSwtInputVerifier on traverse and, if no traverse event was fired, on focusOut
Also adjusted delay to the same value used in swing and rap ui.