removes mail tests from 'default' profile in commons.test fragment
to be able to execute the Mail-Tests javax.mail has to be in the
endorsed dir of the running jre. As this is not the case for every
environment the said tests are excluded in the 'default' profile.

If javax.mail is in the jre classpath the tests can be run with the
profile 'javaxMail' (-P javaxMail).

Change-Id: I0f895335b93d6892a380ba807cc7283d3cec116e