Bug 424917 - Rap: Date chooser dialog opens at wrong location
https://bugs.eclipse.org/bugs/show_bug.cgi?id=424917

Misbehaviour was caused by the text size determination of rap. In
org.eclipse.jface.window.Window#create the method initializeBounds is
called again on a temporary resize which reinitialized the bounds and
overwrote the existing position. Fixed by moving the computation of the
initial location to a method which is used by initializeBounds().

Change-Id: I010e904f30499ab2ebd830734ab55413154005de
Reviewed-on: https://git.eclipse.org/r/20291
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
IP-Clean: Claudio Guglielmo <claudio.guglielmo@bsiag.com>