423933: JEE Server Side Equinox very slow due to complex and repeated class loading (trees)
 - Revert getResource(s) methods but add loop detection.
 - Add property for enabling resource cache analog to the class cache.
 - Add property for enabling usage of bundle order prefix list as
resourceFilter

Resource Cache:
The BundleListClassLoader can cache the resources URLs it finds in the
getResource(s) methods. This has the potential problem, that when
bundles get unloaded, the cached resource URLs can't be found anymore.
Instead of building a big check around this, it was safer to only let it
enable if this should never be the case. 
This method returns the value as it is configured in the config.ini file
or as system property:
org.eclipse.scout.commons.serialization.enableResourceUrlCaching

Resource Filter:
The methods getResource(s) from BundleListClassLoader return resources
from every bundle from the product plus from the general classpath.
There are cases (library bundles with multiple jars) which this behavior
has to be changed. With this property resources will be filtered the
following way:

* if a resource is found in a bundle which matches the prefix list just
this or these resources URLs will be returned.
* if no resource is found in a bundle which matches the prefix list the
list will be returned unfiltered.

This method returns the value as it is configured in the config.ini file
or as system property:
org.eclipse.scout.commons.serialization.enableUsageOfBundleOrderPrefixListAsResourceFilter

Change-Id: I4b3e177da639b099ea438b873dd2400b74c1786f
Task-Url: https://bugs.eclipse.org/bugs/show_bug.cgi?id=423933
Reviewed-on: https://git.eclipse.org/r/27720
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>