HtmlUi: Use "display: table-cell" for table cells
Using "display: table-cell" for table cells (instead of "inline-block")
results in the browser automatically adjusting the heights of all cells
in a row. To work correctly, this needs setting min-width/max-width
instead of width for the cells.

Other than that, we now use "transparent" as background color instead of
"white" for widgets.