HtmlUi: Fixed selected row borders when using browser zoom
Use invisible border that changes color for selected rows instead of a
padding that will be replaced by a border for selected rows) because
'border' and 'padding' behave differently when using browser zoom. By
consistently using borders, we get a mostly consistent behavior.

This solution is not perfect yet, there are issues in the corners of a
row when using a very large zoom (because left and top border have a
different color).

Change-Id: I5cf807cfd2cd78020f2c90016ce6a4a06b8881d6