HtmlUi: Fixed table selection (border corners)
Because the drawing of border corners cannot be influenced, the
appearance is inconsistent across browsers when the left line is blue
and the top line is transparent. We fix that be removing the top border
from the row div and adding a transparent top border to each individual
cell div.

The rendering of the bottom line seems to be different, so this
workaround is not needed there.

Change-Id: Ia9699aaae5f154f29e6f19add743e295bbd9d744