HtmlUi: Use HTML 5 sessionStorage to remember clientSessionId
Instead of encoding it in the URL as query parameter (?clientSessionId)
the clientSessionId is now stored in HTML 5's sessionStorage. This
session storage is inherited by windows opened from within a tab, but
now between different tabs.

The clientSessionId is no longer an ascending number, because it is not
visible anymore to the user. We use the current timestamp with millis
(in base-62 to have a shorter string).

Change-Id: If63fc2cc6599136940439277b3d0ad3677460de0