HtmlUi: Refactoring JSON protocol (type_ => type)
Instead of storing property values under their name in the JSON event
object, they are now wrapped in another JSON object which is available
as 'properties' in the event object. This slightly increases the size of
the JSON message, but with GZIP compression and other optimizations,
this should not be too significant. The advantage is, that this
mechanism works with all possible property names (e.g. 'id') and makes
it unnecessary to use the strange identifier 'type_'. Event attributes
are now always hard-coded, only attributes of the 'properties' events
are dynamic.

Change-Id: I57617e92862d5820e50910d4c56271c1aee2c29c