Bug 440331 - Swing: scheduled client jobs will not be executed on shutdown
https://bugs.eclipse.org/bugs/show_bug.cgi?id=440331

On stopping the client session, wait a short time for all ClientJobs to
complete. This allows for example the SwingScoutRootFrame to store it's
bounds to a service on shutdown. Otherwise setActive(false) would cause
the main thread to end, interrupting any jobs still running.

The maximum time to wait for jobs on shutdown can be customized by using
the setter of the property m_maxShutdownWaitTime.

Additional change: SwingInputVerifier must be executed within
handleSwingWindowClosing() instead of handleScoutDesktopClosedInSwing().
Otherwise, the input verifier could not block closing the application
(which clearly is the code's intention).

Change-Id: I5284c7775146d037cd26c8dcc28598f9d3d90711
Reviewed-on: https://git.eclipse.org/r/30425
Tested-by: Hudson CI
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>