437268: [RAP] Different Display Error Message in Log https://bugs.eclipse.org/bugs/show_bug.cgi?id=437268
Bugfix that closing the desktop due to httpSession invalidation or
userAgent switch was not run on behalf of the Scout model thread but on
the calling thread instead.

Consequences:

(1) The client session was not closed at all because of an early dropout
which was caused due to an unhandled IllegalStateException when queuing
into the UIThread.

(2) When switching the userAgent, the cleanup was run on behalf of the
new UIThread with its own Display registered. In turn, the Display of
the current Thread differed from the one registered within the
environment.

Change-Id: I901f4c33b992cb6b0020ed8bc4440c6c0e298195
Reviewed-on: https://git.eclipse.org/r/32627
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>