HtmlUi: Improved table row order changed event
- Added code which prevents animation for newly inserted rows. This is
necessary since the new rows are always added at the end in the scout
table model. The ordering is done afterwards. This leads to a movement
of the new row at the last position to a new position right after
insertion which is very confusing. The animation is now only done for
already existing rows.