Bug 444427 - JVM Fatal Crash because of IE Patch KB2977629:
Fix to solve the problem in Swing UI, that after disposing the browser
widget, some browser-events still might be consumed by the SWT event
dispatcher which caused the JVM to crash (consequence of KB2977629). To
bypass the Fatal error, the Display is disposed instead of the browser
widget. Thereto a browser field must have its dedicated Display.

Change-Id: Iba1b72fcf36568034fc15ca31fc911358eff0e8f
Reviewed-on: https://git.eclipse.org/r/33568
Tested-by: Hudson CI
Reviewed-by: André Wegmüller <awe@bsiag.com>
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsiag.com>
Reviewed-on: https://git.eclipse.org/r/33596