Bug 440331 (additional fixes) - Swing: scheduled client jobs will not be executed on shutdown
https://bugs.eclipse.org/bugs/show_bug.cgi?id=440331

This fixes an additional issue when closing the application. Because the
server-side logout was performed before all client jobs were given the
chance to finish, it was possible that the server session was always
re-created on closing the client. To prevent this, the call to the
logout service (which invalidates the HTTP session) must be deferred
until the waiting period for client jobs has expired. Note that a
"ClientAsyncJob" has to be used instead of simple "Job", because
otherwise, the ILogoutService could not be resolved.

Additional improvements:
- Make private method scheduleSessionInactivation() protected
- Remove unnecessary "final" modifiers
- ClientNotificationPollingJob: Additional check if job was cancelled

Change-Id: I811ddee9b9a4ea004d78dae4ffac0be87f1fc344
Reviewed-on: https://git.eclipse.org/r/33791
Tested-by: Hudson CI
Reviewed-by: Christoph Schwank <christoph.schwank@bsiag.com>
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>
Reviewed-by: Beat Schwarzentrub <bsh@bsiag.com>