Bug #446208: Missing visual marker for editable cells in RAP UI.
The fix is mainly done by externally patching RAP 'GridRow.js'. This is
necessary because there is no variant-support for individual table
cells. Alternatively, an additional HTML-element could be introduced to
wrap the cell's content. However, this causes some performance overhead
and a hack is still requird to reset the cell's padding because being
calculated directly into the absolute cell position.

Change-Id: Ia168cb5622fe4df8317fc833bab624b8889d69c3
Reviewed-on: https://git.eclipse.org/r/34640
Tested-by: Hudson CI
Reviewed-by: Claudio Guglielmo <claudio.guglielmo@bsiag.com>