Bug 443555/Bug 441045: SWT: Smartfield does not work in Linux and Mac OS X https://bugs.eclipse.org/bugs/show_bug.cgi?id=443555 https://bugs.eclipse.org/bugs/show_bug.cgi?id=441045
Fixed the various problems when requesting the proposal popup and
accepting proposals in form-fields as well as in cell-editors. The main
issue was that the popup dialog became the active shell and focus owner
which triggered to many verify events. Furthermore, the event and focus
handling of cell-editors that open a Shell was facilitated and improved
to have the same behavior as in a smartfield cell editor.


Change-Id: I880a86c0dbc7cb78af21ed8b40061d21f5fbef6d
Reviewed-on: https://git.eclipse.org/r/35219
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsiag.com>
Tested-by: Hudson CI
Reviewed-on: https://git.eclipse.org/r/40786
Reviewed-by: Andreas Hoegger <Andreas.Hoegger@bsiag.com>