Bug 453504 - SWT: Nullpointer when toggling a checkable table-cell https://bugs.eclipse.org/bugs/show_bug.cgi?id=453504
The problem was based that a traversal event was fired the time the
cell-editor was focused. As having no focusable control at all and being
closed immediately, that resulted in accessing a 'null' cell-editor
reference.
Furthermore, simplified the usage of cell-editors by using JFace editing
support with a separate cell-editor for each cell.

Change-Id: Ieac92dac605283dcdc26735922374b24588469dd
Reviewed-on: https://git.eclipse.org/r/37287
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsiag.com>