HtmlUi: Improved json adapter disposal
- Every adapter now has exactly one parent
- There may be multiple adapters for one model instance
- If an adapter should be shared amongts adapters, one has to create a
global adapter. This is currently done for forms, messageboxes,
outlines, detail tables
- Child adapters now get disposed on property change (a global adapter
is not a child -> does not get disposed)
- Child adapters now get disposed when the parent gets disposed -> no
need to implement disposeChildAdapters anymore

This makes sure that adapters get disposed when not needed anymore with
the benefit of having shared (global) adapters to improve performance.

Another solution would have been to keep track of the usage of an
adapter. But that solution has the drawback that either the same logic
needs to be implemented on the client or that the server needs to send
dispose events for every object -> Current solution is more simpler and
should perform better (no dispose events necessary).