HtmlUi: Implemented disposal on client side
- Implemented same mechanism as on server
- Parent destroys its children -> Global adapters don't get destroyed
- There is a root adapter which holds the global adapters. The root
adapter is created on client side and always has id '1'.
- Adapters get destroyed on property change events
- The information, whether the adapter is global, is sent by the server.
This is: adapter.parent: rootAdapter.id.
- The reference to the parent is only sent in case of a root adapter
reference. In every other case the parent is the creator of the adapter
which has to correlate with the server. 
Note: At the moment we think this is always true. If this turns out to
be wrong, we'll have to always send the parent id.