SWT: Accept focus requests even if the workbench is blocked https://bugs.eclipse.org/bugs/show_bug.cgi?id=457859
Introduced a mechanism of being notified once the application exits the
blocking mode.

SWT specific:
The time the workbench is blocked, focus requests are ignored because
the workbench filters the resulting OS-event from the SWT event queue.
That is why focus requests can only be applied if long-running
operations complete.

For more information see the forum thread:
https://www.eclipse.org/forums/index.php/t/759748/
Change-Id: I5fc6ca5a4f3a0e960cf514511761a3b889b79a04
Reviewed-on: https://git.eclipse.org/r/39906
Tested-by: Hudson CI
Reviewed-by: Judith Gull <jgu@bsiag.com>