HtmlUi: Fixed intial display text for value field
Initially, the display text is not set. This should be equal to an empty
displaytext, because this cannot be distinguied by the user ("no
content" and "empty content" has to same representation in the UI
field). If a field is left unchanged, there should be no change event
from undefined to ''.