HtmlUi: Split parent relation into parent & owner relation
Owner = Parent model adapter from server (outer adapter or the global
root adapter). This relation is relevant when creating and destroying an
adapter.
Parent = Outer adapter where the current adapter is rendered in. This
may be different from the owner, e.g. when the owner is the global root
adapter. For example a Form can have owner=GlobalAdapter and
parent=WrappedFormField.

Accordingly, the children list was split in ownedAdapters and children.