HtmlUi: Added possibility to coalesce gui events
Now, if a table column gets resized, only the last event gets sent. This
was not necessary so far, because the column could only be resized by
using the column seperator where we had the possibility to track the
resizeEnd. With the introduction of autoResizeColumns, resizing the
columns may be triggered by resizing the window where it is not possible
to get a resizeEnd. 
However, sending resize events is still disabled if autoResizeColumns is
set to true. Just because it is of no use mostly.