HtmlUi: Apply buffered events before disposing nodes and rows
This fixes "leak detection" errors that may occur in the following
situation: When the JSON adapter of a table or tree is disposed, we
query the model for all remaining rows or nodes and dispose them. After
that, no mappings should remain. However, when the model already removed
all rows or nodes, but the corresponding model events are not applied to
the JSON adapter (because of the eventBuffer), the dispose method sees
no model items anymore, but still finds internal mappings. Therefore,
the event buffer has to be flushed before handling the disposal.