HtmlUi: Replace wait/notify by notification blocking queue
This improves the notification of background job requests. The plain
Java wait/notifiy mechanism cannot correctly handle the case when the
poller is currently not waiting on the server. The notification would be
missed. By replacing the mechanism by a blocking queue with a
"notification token", the notification will remain in the memory for the
next polling request.