HtmlUi: Fixed exception when removing a form
UseCase: DetailForm was open, user clicks edit and saves the form. The
save triggers a reload of the table -> detail form gets removed. Because
the detailform had no link to the tab, removeForm only removed the form
but not the tab -> tab still had a link to form.$container with
tab.$storage. But tab.$storage was removed (due to form.remove()) and
therefore did not contain any data anymore -> HtmlComponent and
DetachHelper threw exceptions.