HtmlUi: Added support for empty table rows and emtpy top labels
- Min-Height would be an alternative to nbsp, the drawback is that it
can't be computed properly so we would have to adjust it whenever the
css changes -> nbsp is always right.
- Use case for top labels: Two fields beside each other, the left one
has a top label, the right one has an empty top label -> the fields
itself are aligned only if the empty top label is as height as the non
empty one.