HtmlUi: Don't send node events too early to the UI
Ignore nodes that are not (yet) sent to the UI. This may happen due to
the asynchronous handling of events, when the model is already up to
date.

Example:
- Tree with 1 Node below the invisible root node.
- expandAllNodesRecursively()
- Insert new node below the first node.

When the event "expand_recursive" is handeled, the new node already
exists in the model, which was not the case when the event was fired. If
we would create a node id for this node here, we would get an "expanded"
event before we get the corresponding "inserted" event.