HtmlUi: resources may specifiy additional HTTP headers
Resources can provide additional HTTP headers. This is useful for
setting the "Content-disposition" header when downloading
BinaryResources. This header should be set for most, but not all
BinaryResources (e.g. not for inline images in the ImageField).

Additionally the X-UA-Compatible <meta> tag was replaced by the
corresponding HTTP header (makes index.html simpler).