HtmlUi: Fix use of json.optString()
optString() returns "" when value is null. In most cases, this is not
correct in Scout, therefore the explicit return value "null" should be
specified.

In some cases, optString() was only used because it does not declare a
checked exception. This is no longer the case, so getString() should be
used instead.