Set row state to UPDATED when row is checked
When a row is checked, it is added to a list of checked rows. If it has
a "checkable column", the value for that column is updated, which causes
a ROWS_UPDATED event to be emitted. However, if no such column is
present, the row state was not changed, and so the table was considered
unchanged. This change manually sets the row state to "updated", but
without firing any ROWS_UPDATED events. The dedicated ROWS_CHECKED event
is used instead.