Fix issue that toggling the visibility of elements in 'not-selected' tabs does not result in a relayout of the parent container https://bugs.eclipse.org/bugs/show_bug.cgi?id=470458
When the visibility of a field on a tab is changed to false the field
disappears and space that was occupied by the field is correctly claimed
by other subsequent components.
When the same tab is inactive at the time the visibility is changed, the
field disappears but the space is not claimed by the other components.
(cherry picked from commit b9929f68340c0251107d2e98a6ace0127ecebb0f)

Change-Id: Ie73fb4084cbfd35a149e98e9c0d72aecc1404ae8
Reviewed-on: https://git.eclipse.org/r/50408
Tested-by: Hudson CI
Reviewed-by: Daniel Wiehl <daniel.wiehl@bsiag.com>