Fix issue that toggling the visibility of elements in 'not-selected' tabs does not result in a relayout of the parent container https://bugs.eclipse.org/bugs/show_bug.cgi?id=470458
When the visibility of a field on a tab is changed to false the field
disappears and space that was occupied by the field is correctly claimed
by other subsequent components.
When the same tab is inactive at the time the visibility is changed, the
field disappears but the space is not claimed by the other components.