TableEventBuffer: Check "same row" by reference instead of rowIndex
To check if two rows are "the same row" in multiple events, it is not
sufficient to check if the row index is the same. When a row is deleted,
it keeps it last row index, but another row gets the same row index.

Example:
[rowA, index 0]
[rowB, index 1]
[rowC, index 2]
[rowD, index 3]

When rowB is deleted, rowC gets the index 1, rowD index 2. When rowC is
also deleted, rowD gets the index 2. But in the event buffer, we have
two events with the same row index!