HtmlUi: Improve rendering of menubar
When the menubar items are rebuilt, the artificial "ellipsis" menu
created by the menubar layout has to be removed as well. Otherwise, it
can be visible at the beginning of the menubar for a short time.

To correctly determine if menu items have to be rebuilt, the original
item list has to be remembered and compared (because the "menuItem" list
may be sorted differently by the menuSorter and contain artificial
separators).

MenuBarKeyStrokeAdapter has been simplified.