Put list of changed columns to ROWS_UPDATED table events
When a cell in a table changes it values, a ROWS_UDPATED event is
triggered for that row. To improve handling of such events, a list of
columns that have been updated is now provided by the event. The handler
therefore gets more information about which parts of the table row have
been updated.