HtmlUi: z-index cleanup
Install a local stacking context for each child element of the .scout
DIV, by explicitly setting the z-index to 0. Elements that should be
opened "on top of everything" should add themselves to the .scout DIV as
popup. Inside of elements, z-index may (but generally should not) be
used. Most occurences use a z-index of 1 or 2 to draw over other
elements. (Exception: key boxes with z-index 100.) Usually, the order
may be specified by inserting the element at the end of the DOM subtree.