HtmlUi: Only update menubar layout when properties have really changed
When e.g. the "enabled" property of a menu changes from TRUE to FALSE
and back to TRUE, this will emit a propertyChanged event from the model,
but the property has not really changed for the UI. Therefore, we check
with the old property values to detect properties that have really
changed.

This fixes layout issues with the ProcessAssistantToolButton when a menu
is clicked.