HtmlUi: Improve URL redirect in login.js
After the authentication was completed sucessfully, the login form
reloads the page to start up the application. Instead of simply
reloading the page, the current location is read and everything after
that last '/' character is removed. If the current URL is
'http://myserver/myapp/login.html', the redirect URL is
'http://myserver/myapp/'. This is useful to explicitly go to the
login.html to force user/password authentication.

Additionally, scout.login.init() has an options argument that allows
setting the redirect URL manually from outside.