HtmlUi: Fix form layout in popup / change HtmlComponent.isAttached()
When a form tool button is unselected, the form is detached. If the
layout becomes invalid in the same request, it cannot be validated
correctly (because the elements are no longer in the DOM). Therefore the
layout has to be validated when the form is re-attached.

The function isRemoved() was removed from HtmlComponent, because it
always returned false (this.$comp.data('htmlComponent') is set in
constructor and never unset). Instead, the method isAttached() was added
that uses $.fn.isAttached(). This method seems to be the fastest method
to check if an element is removed or detached. See discussion at
http://stackoverflow.com/questions/3086068