HtmlUi: Move parsing from AbstractDateField to the UI
AbstractDateField no longer parses input to determine a date. Instead,
parsing (and formatting) is done entirely in the UI. The model gets an
already validated date object as new value.

To fix timezone issues between UI and server, the "JSON date" format is
used to transfer the timestamp, not a millisecond value.

Change-Id: If7b7995ed2b7ecb8c1bcaf03e38379bd35faad38