HtmlUi: Move parsing from AbstractDateField to the UI
AbstractDateField no longer parses input to determine a date. Instead,
parsing (and formatting) is done entirely in the UI. The model gets an
already validated date object as new value.

DateFormat.js and DateField.js were reworked almost entirely to fix
various issues in date and time parsing and prediction.

To fix timezone issues between UI and server, the "JSON date" format is
used to transfer the timestamp, not a millisecond value.

Change-Id: Icc5e191e9dbbb62fc0c12d93b0beaade9345011e