WrappedFormField: Don't remove m_innerForm when disposing the field
When the AbstractWrappedFormField is disposed, it should remove all
listeners from the inner form and (if manageFormLifeCycle=true) close
it. However, because some values of the inner form might be read later,
the reference to the inner form must not be removed - even if the inner
form is already closed!