HtmlUi: Fix JS error in ModalityController when showing fatal message
Nothing to be done if rendering the classpane is manually disabled by
setting "render" to false (see Form.js) or there is no displayParent.
The later case may happen while showing a popup before the desktop is
available (e.g. fatal message). When no desktop is present, a glasspane
is not necessary.