HtmlUi: Fix long "busy" dialog after reconnect / improved offline msg
Bugfix: When the session is offline and becomes online again, the busy
dialog appeared for "a long time". This was caused by the
"pollForBackgroundJobs" request to be added to the queuedRequest. This
special request has to be ignored.

Improvement: A very basic implementation of "notifications" (as
suggested by the designers) was added to BaseDesktop, offline-message
was changed to look more like a notification. Note: This feature needs
more work.