Make desktop-based form events to be transported by Desktop adapter
The motivation is that preceding events are sallowed once an adapter is destroyed. That causes the events not to be sent to the UI. As a consequence, forms might not be hidden, e.g. when the display parent is changed, which results in an 'already rendered widget' exception.

Review: claudio.guglielmo@bsi-software.com