HtmlUi: Table must not flicker when pressing tab for cell editing
CompleteCellEdit now does not remove the editor immediately anymore, it
is now done by endCellEdit. This has the advantage, that closing the old
and opening the new happens at the same time -> no flickering visible.
Another advantage: One cannot see the old value in the table cell when
pressing tab. Previously it was visible shortly which does not look
well.