HtmlUi: Fix menubar "unrendered last item problem"
When the menubar is removed, all interal data structures have to be
emptied as well. Otherwise, the old menu items might still be present,
but are not rendered. When they should be rendered, they will always be
recreated again. Instead of leaving the old items and checking if they
are rendered (and ignore them if not), we remove them completely. Like
this, no rendered checks are necessary.