HtmlUi: Session.goOffline() should propagate to rootAdapter, not desktop
When the session goes offline or online, this event should be propagated
to the "root adapter". The desktop is owned by the root adapter, will
therefore be notified eventually. But there are other components (e.g.
outlines) that are not owned by the desktop and would miss the
offline/online event.