HtmlUi: Replace formClosed event by 'disposeAdapter' event to session
When a form is closed, it has to be explicitly disposed on the UI,
because the forms parent is not disposed. Using the 'formClosed' event
on the form itself is a hack, because usually events for disposed
adapters are removed form the JSON response. Even worse, when the form
is opened and closed in the same request, an error occurs on the UI,
because it tries to handle an event for an adapter it had never
received!

To fix this issue, the 'formClosed' event on the form itself was
removed. Instead, a 'disposeAdapter' adapter event was introduced on the
UI session, that may be used to explicitly dispose a adapter on the UI.
(Currently only necessary for forms, because all other adapter are
disposed automatically when their parent is disposed.)