HtmlUi: Prevent accidental execution of context menu
Menus should only react to left mouse button, not to other buttons. This
also fixes the following bug: when the right mouse button is clicked and
a context menu is opened, it could happen that one of the menus as
executed immediately, because the context menu popup was rendered under
the mouse cursor (e.g. on the right border of the screen). When the
right mouse button was released, apparently a click event was generated.
By ignoring events for other buttons than the left mouse button, the
described bug is fixed automatically.