Propagate "table popuplate status" as property to ITable
In some cases, no table field is present to carry the table popuplate
status to the UI (e.g. in desktop tables). Therefore, the table
popuplate status is propagated from the page and the table field to the
table. The "table selection status" is ignored, because the Html UI
renders selection by itself.

For "auto sum" functionality, a solution is still to find!