Fix issue with popups used in CellEditorPopup
This fix makes smartfield or datefield popups not to fire 'popupopen' event to not close other popups. That is crucial if being used within other popups like CellEditorPopup. Otherwise, that 'parent' popup would be closed immediately if opening a smart- or datefield popup.

Also fix issue that datefield might be disposed/removed once handling a keystroke.