HtmlUi: HtmlDocumentParser: Trim replacement before adding it to result
Because the tags in a HTML file may not necessarily start at the
beginning of a line, we add a newline before the replacement. The
replacement file itself may contain whitespace at the end. This is
stripped and replaced with a single newline.

Additionally, a typo was fixed (inlcudeUrl) and TEXTS.get() fallback was
removed, because that is already provided by TESTS.get().