HtmlUi: Fix concurrency issue with m_jsonAdapterRegistry
Access to m_jsonAdapterRegistry has to be synchronized. This is usually
ensured by the uiSessionLock. However, if a servlet request is in
processJsonRequest() (but not in a model job) and a model job is started
(asynchronous by the model, e.g. from client notifications), the access
is not synchronized. The model job does not get the uiSessionLock.
Therefore, the adapter registry must not be accessed from outside a
model job!

To fix the issue, the debug messages were moved inside a model job.

#162449