HtmlUi: Immediately dispose deleted nodes
This fixes issues with events being processed on tree nodes that are
already deleted. 

Example: Assume a table page A with another table page B as a child.
When due to some event in the same model job first the page B reloads
all its children and then node A, there would be the following events:
- delete all rows in table B
- insert new rows in table B
- delete all rows in table A
- delete all child nodes of node A
- insert new rows in table A
- insert new child nodes in node A
When processing the first two events, the corresponding node B would
already be detached from its tree. JsonOutline wants to use this tree to
find the corresponding nodeId. This will therefore fail.

As a fix, when a node is deleted, the nodes are disposed immediately.
This will cause the corresponding table to be disposed, and with it the
"wrong" events. They will never be processed (which is correct, as they
are obsolete).